deploy/ansible/requirements.txt
ansible-compat==24.6.1
ansible-core==2.16.0
ansible-lint==24.6.1
attrs==25.1.0
azure-appconfiguration==1.7.1
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
black==25.1.0
bracex==2.5.post1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
filelock==3.17.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.31.1
msal-extensions==1.2.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
portalocker==2.10.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
resolvelib==1.0.1
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
six==1.17.0
subprocess-tee==0.4.2
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
wcmatch==10.0
yamllint==1.35.1
zipp==3.21.0
